backend/requirements.txt
uvicorn==0.29.0
fastapi==0.110.0
python-dotenv==1.0.1
httpx==0.27.0
qdrant-client==1.8.0
sentence-transformers
pytest==8.2.2
pytest-asyncio==0.23.7